refactor(#288): split useStreamEvents.ts into event-type handlers
