+++ apps/desktop/src/hooks/streamEventHandlers/index.ts
@@ -1,0 +1,4 @@
+export { handleToolEvent } from './toolEventHandlers';
+export { handleMessageEvent } from './messageEventHandlers';
+export { handleSystemEvent } from './systemEventHandlers';
+export { handlePermissionPlanEvent } from './permissionPlanHandlers';
